Add pagination logic on backend and frontend
* Manage pagination on backend. Add validations to handle when
this attribute is not given.
* Add second variable on Sprintf function to add a variable
to from query attribute.
* Move getEmails function to a different folder to make it
accessible to all folders. Also, modify it to handle the page
attribute.
* Add conditions to show results when the page changes and when
is not change.
* Add two new methods to the store to manage to remove the email
select and to manage when to show the loader.
* Add loader asset to show that the request is on load.
* Add parameter to getEmails to know when the page is in use and
condition the results to be save on the store.
* Add button to load more emails. This button fire the getEmails
function to get more emails.
* Add condition on submit function to clean the emails and the
email select when the term is erase from the input.